Move addApplicationNote service method to submittedApplication service
Seeing as this is an operation on a submitted application, this is a more suitable location.
